+++ src/solanum/output/riemann.clj
@@ -44,16 +44,17 @@
 
   (write-events
     [this events]
+    (when-not (riemann/connected? client)
+      (riemann/reconnect! client))
     @(riemann/send-events client (prepare-batch events))))
 
 
 (defmethod output/initialize :riemann
   [config]
-  (let [host (get config :host "localhost")
-        port (get config :port 5555)]
+  (let [host (str (get config :host "localhost"))
+        port (int (get config :port 5555))]
     (map->RiemannOutput
       {:type :riemann
        :host host
        :port port
-       :client (doto (RiemannClient/tcp (str host) (int port))
-                 (riemann/connect!))})))
+       :client (RiemannClient/tcp host port)})))